Budget_101/requirements.txt
streamlit>=1.26.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.8.0
pillow>=9.0.0
numpy>=1.20.0

